+++ packages/streamr-logger/include/streamr-logger/Logger.hpp
@@ -1,29 +1,14 @@
 #ifndef STREAMER_UTILS_LOGGER_LOGGER_H
 #define STREAMER_UTILS_LOGGER_LOGGER_H
 
-#include <iostream>
-#include <memory>
-#include <folly/Format.h>
-#include <folly/detail/StaticSingletonManager.h>
-#include <folly/logging/FileHandlerFactory.h>
-#include <folly/logging/LogCategory.h>
-#include <folly/logging/LogCategoryConfig.h>
 #include <folly/logging/LogConfig.h>
-#include <folly/logging/LogConfigParser.h>
-#include <folly/logging/LogFormatter.h>
-#include <folly/logging/LogHandler.h>
-#include <folly/logging/LogHandlerConfig.h>
 #include <folly/logging/LogLevel.h>
 #include <folly/logging/LogMessage.h>
 #include <folly/logging/LogWriter.h>
-#include <folly/logging/Logger.h>
 #include <folly/logging/LoggerDB.h>
-#include <folly/logging/StandardLogHandler.h>
-#include <folly/logging/StandardLogHandlerFactory.h>
-#include <folly/logging/StreamHandlerFactory.h>
 #include <folly/logging/xlog.h>
-#include <folly/portability/Time.h>
 #include "StreamrWriterFactory.hpp"
+#include "StreamrHandlerFactory.hpp"
 
 namespace streamr {
 namespace logger {
@@ -48,227 +33,25 @@ const std::unordered_map<std::string, folly::LogLevel> toFollyLogLevelMap{
     {logLevelErrorText, folly::LogLevel::ERR},
     {logLevelFatalText, folly::LogLevel::CRITICAL}};
 
-struct StreamrLogMessage {
-    std::chrono::system_clock::time_point timestamp;
-    folly::StringPiece fileBasename;
-    unsigned int lineNumber;
-    folly::LogLevel logLevel;
-    const std::string& logMessage;
-};
+std::optional<folly::LogLevel> getFollyLogLevelFromEnv() {
+    char* val = getenv(envLogLevelName);
+    if (val) {
+        auto pair = toFollyLogLevelMap.find(val);
+        if (pair != toFollyLogLevelMap.end()) {
+            return std::optional<folly::LogLevel>{std::move(pair->second)};
+        }
+    }
+    return std::nullopt;
+}
 
 }; // namespace
-/*
-Format based on this Streamr log. The Filename and line number is fixed size
-with 36 characters. If it is longer then it is truncated
-DEBUG [2024-06-06T13:52:49.816] (<Filename>: <Line Number>     ):<Message>
-
-Examples of a truncated filename
-ERROR [2024-06-06T13:52:49.816] (12345678901234567890123456789012: 10):<Message>
-ERROR [2024-06-06T13:52:49.816] (1234567890123456789012345678901: 101):<Message>
-
-This one is not truncated:
-ERROR [2024-06-06T13:52:49.816] (1234567890123456.cpp: 101           ):<Message>
-
-DEBUG [2024-06-05T08:50:39.787] (File name): Message
-ERROR [2024-06-05T08:50:39.787] (File name): Message
-FATAL [2024-06-05T08:50:39.787] (File name): Message
-INFO [2024-06-05T08:50:39.787] (File name): Message
-TRACE [2024-06-05T08:50:39.787] (File name): Message
-WARN [2024-06-05T08:50:39.787] (File name): Message
-*/
-/*
-class StreamrWriterFactory : public folly::StreamHandlerFactory::WriterFactory {
-   public:
-    StreamrWriterFactory() : logWriter_{nullptr} {}
-
-    StreamrWriterFactory(std::shared_ptr<folly::LogWriter> logWriter)
-        : logWriter_{logWriter} {}
-
-    std::shared_ptr<folly::LogWriter> createWriter() override {
-        if (logWriter_) {
-            return logWriter_;
-        } else {
-            return folly::StreamHandlerFactory::WriterFactory::createWriter();
-        }
-    }
-
-   private:
-    std::shared_ptr<folly::LogWriter> logWriter_ = nullptr;
-};
-*/
-class StreamrLogFormatter : public folly::LogFormatter {
-   public:
-    StreamrLogFormatter() = default;
-    ~StreamrLogFormatter() override = default;
-
-    struct StreamrLogMessage {
-        std::chrono::system_clock::time_point timestamp;
-        folly::StringPiece fileBasename;
-        unsigned int lineNumber;
-        folly::LogLevel logLevel;
-        const std::string& logMessage;
-    };
-
-    std::string formatMessageInStreamrStyle(const StreamrLogMessage& message) {
-        struct tm ltime;
-        const auto timeSinceEpoch = message.timestamp.time_since_epoch();
-        const auto epochSeconds =
-            std::chrono::duration_cast<std::chrono::seconds>(timeSinceEpoch);
-        const std::chrono::milliseconds millisecs =
-            std::chrono::duration_cast<std::chrono::milliseconds>(
-                timeSinceEpoch) -
-            epochSeconds;
-        time_t unixTimestamp = epochSeconds.count();
-        if (!localtime_r(&unixTimestamp, &ltime)) {
-            memset(&ltime, 0, sizeof(ltime));
-        }
-        auto basename = message.fileBasename;
-        const auto fileNameLength = std::ssize(basename);
-        const auto lineNumberInString = std::to_string(message.lineNumber);
-        const auto lineNumberLength = std::ssize(lineNumberInString);
-        const std::string logMessageColor = "\033[36m";
-        const std::string logMessageColorReset = "\033[0m";
-        const auto fileNameAndLineNumberLength =
-            (fileNameLength + lineNumberLength + separatorLength);
-        // Is filename is truncated if filename, separator and lineNumber does
-        // not fit to the fixed size
-        if (fileNameAndLineNumberLength <= maxFileNameAndLineNumberLength) {
-            basename = basename.toString()
-                           .append(fileNameAndLineNumberSeparator)
-                           .append(lineNumberInString);
-            auto logLine = folly::sformat(
-                "{}{}{} [{:04d}-{:02d}-{:02d}T{:02d}:{:02d}:{:02d}.{}] ({:<36}): {}{}{}\n",
-                getColorSequence(message.logLevel),
-                getLogLevelName(message.logLevel),
-                logMessageColorReset,
-                ltime.tm_year + 1900,
-                ltime.tm_mon + 1,
-                ltime.tm_mday,
-                ltime.tm_hour,
-                ltime.tm_min,
-                ltime.tm_sec,
-                millisecs.count(),
-                basename,
-                logMessageColor,
-                message.logMessage,
-                logMessageColorReset);
-            return logLine;
-        } else {
-            // Truncate needed
-            auto lengthForTruncatedFileName = maxFileNameAndLineNumberLength -
-                (lineNumberLength + separatorLength);
-            basename = basename.substr(0, lengthForTruncatedFileName);
-            auto logLine = folly::sformat(
-                "{}{}{} [{:04d}-{:02d}-{:02d}T{:02d}:{:02d}:{:02d}.{}] ({: <*}{}{}): {}{}{}\n",
-                getColorSequence(message.logLevel),
-                getLogLevelName(message.logLevel),
-                logMessageColorReset,
-                ltime.tm_year + 1900,
-                ltime.tm_mon + 1,
-                ltime.tm_mday,
-                ltime.tm_hour,
-                ltime.tm_min,
-                ltime.tm_sec,
-                millisecs.count(),
-                lengthForTruncatedFileName,
-                basename,
-                fileNameAndLineNumberSeparator,
-                lineNumberInString,
-                logMessageColor,
-                message.logMessage,
-                logMessageColorReset);
-            return logLine;
-        }
-    }
-
-    std::string formatMessage(
-        const folly::LogMessage& message,
-        const folly::LogCategory* handlerCategory) override {
-        return this->formatMessageInStreamrStyle(
-            {message.getTimestamp(),
-             message.getFileBaseName(),
-             message.getLineNumber(),
-             message.getLevel(),
-             message.getMessage()});
-    }
-
-   private:
-    static constexpr int maxFileNameAndLineNumberLength{36};
-    static constexpr folly::StringPiece fileNameAndLineNumberSeparator{": "};
-    static constexpr auto separatorLength{
-        std::ssize(fileNameAndLineNumberSeparator)};
-
-    constexpr folly::StringPiece getLogLevelName(folly::LogLevel level) {
-        if (level == folly::LogLevel::DBG) {
-            return "TRACE";
-        } else if (level == folly::LogLevel::DBG0) {
-            return "DEBUG";
-        } else if (level == folly::LogLevel::INFO) {
-            return "INFO";
-        } else if (level == folly::LogLevel::WARN) {
-            return "WARN";
-        } else if (level == folly::LogLevel::ERR) {
-            return "ERROR";
-        }
-        return "FATAL";
-    }
-
-    folly::StringPiece getColorSequence(folly::LogLevel level) {
-        if (level == folly::LogLevel::DBG) {
-            return "\033[90m"; // Gray (TRACE)
-        } else if (level == folly::LogLevel::DBG0) {
-            return "\033[34m"; // Blue (DEBUG)
-        } else if (level == folly::LogLevel::INFO) {
-            return "\033[32m"; // Green (INFO)
-        } else if (level == folly::LogLevel::WARN) {
-            return "\033[33m"; // Yellow (WARN)
-        } else if (level == folly::LogLevel::ERR) {
-            return "\033[31m"; // Red (ERROR)
-        }
-        return "\033[1;41m"; // Red Background (FATAL)
-    }
-};
-
-class StreamrLogFormatterFactory
-    : public folly::StandardLogHandlerFactory::FormatterFactory {
-   public:
-    bool processOption(
-        folly::StringPiece name, folly::StringPiece value) override {
-        return false;
-    }
-
-    std::shared_ptr<folly::LogFormatter> createFormatter(
-        const std::shared_ptr<folly::LogWriter>&) override {
-        return std::make_shared<StreamrLogFormatter>();
-    }
-};
-
-class StreamrHandlerFactory : public folly::StreamHandlerFactory {
-   public:
-    StreamrHandlerFactory() = default;
-    ~StreamrHandlerFactory() override = default;
-    StreamrHandlerFactory(StreamrWriterFactory* writerFactory)
-        : writerFactory_{writerFactory} {}
-
-    std::shared_ptr<folly::LogHandler> createHandler(
-        const Options& options) override {
-        //  StreamHandlerFactory::WriterFactory writerFactory;
-        StreamrLogFormatterFactory formatterFactory;
-
-        return folly::StandardLogHandlerFactory::createHandler(
-            getType(), writerFactory_, &formatterFactory, options);
-    }
-
-   private:
-    StreamrWriterFactory* writerFactory_;
-};
 
 class Logger {
-   public:
+public:
+
     Logger(std::shared_ptr<folly::LogWriter> logWriter = nullptr)
         : loggerDB_{folly::LoggerDB::get()} {
         if (logWriter) {
-            // logWriter_ = std::make_shared<LogWriterMock>();
             writerFactory_ = StreamrWriterFactory(logWriter);
         } else {
             writerFactory_ = StreamrWriterFactory();
@@ -283,20 +66,6 @@ class Logger {
         return instance;
     }
 
-    // Used only for unit testing
-    Logger(
-        folly::LoggerDB& loggerDB,
-        std::shared_ptr<folly::LogWriter> logWriter,
-        bool isInitialized = true)
-        : loggerDB_{loggerDB}, logWriter_{logWriter} {
-        if (isInitialized) {
-            writerFactory_ = StreamrWriterFactory(logWriter_);
-            logHandlerFactory_ = {
-                std::make_unique<StreamrHandlerFactory>(&writerFactory_)};
-            this->initializeLoggerDB(folly::LogLevel::DBG);
-        }
-    }
-
     void log(folly::LogLevel level, const std::string& message) {
         auto follyLogLevel = getFollyLogLevelFromEnv();
         if (follyLogLevel) {
@@ -304,7 +73,6 @@ class Logger {
                 // loggerDB.setLevel("", *follyLogLevel);
                 this->initializeLoggerDB(*follyLogLevel, true);
             }
-
             // This complicated code is mostly copy/pasted from folly's XLOG
             // macro. It is not very flexible to call macro from here so. This
             // code below prints a log message without any return code.
@@ -331,25 +99,13 @@ class Logger {
         }
     }
 
-   private:
+private:
     std::shared_ptr<folly::LogWriter> logWriter_;
     StreamrWriterFactory writerFactory_;
     folly::LoggerDB& loggerDB_;
     std::unique_ptr<folly::LogHandlerFactory> logHandlerFactory_;
-    // folly::StreamHandlerFactory::WriterFactory writerFactory_;
-    std::optional<folly::LogLevel> getFollyLogLevelFromEnv() {
-        char* val = getenv(envLogLevelName);
-        if (val) {
-            auto pair = toFollyLogLevelMap.find(val);
-            if (pair != toFollyLogLevelMap.end()) {
-                return std::optional<folly::LogLevel>{std::move(pair->second)};
-            }
-        }
-        return std::nullopt;
-    }
-
+   
     void initializeLoggerDB(
-        //  folly::LoggerDB& db,
         const folly::LogLevel& rootLogLevel,
         bool isSkipRegisterHandlerFactory = false) {
         if (!isSkipRegisterHandlerFactory) {
